change file location to SearchRecipes module add shadow layer to image and fix the design and image
